services/tenant-api/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic>=2.9,<3
httpx>=0.27,<1
PyJWT[crypto]>=2.8,<3
